project/project/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
crypto==1.4.1
distlib==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
django-localflavor==3.1
django-phone-field==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.2
idna==3.4
importlib-metadata==5.1.0
Naked==0.1.32
Pillow==9.3.0
platformdirs==2.6.0
pycryptodome==3.16.0
PyJWT==2.6.0
python-stdnum==1.18
pytz==2022.6
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
zipp==3.11.0
